+++ program_management/templates/search.html
@@ -132,9 +132,13 @@
                 <input type="hidden" id="hdn_with_activities_organization" name="with_activities_organization">
                 <input type="hidden" id="hdn_with_responsibles_and_contacts" name="with_responsibles_and_contacts">
                 <input type="hidden" id="hdn_with_diplomas_and_certificats" name="with_diplomas_and_certificats">
-
-
-
+                <input type="hidden" id="hdn_with_co_graduation_and_partnership" name="with_co_graduation_and_partnership">
+                <input type="hidden" id="hdn_with_ares_code" name="with_ares_code">
+                <input type="hidden" id="hdn_with_enrollment" name="with_enrollment">
+                <input type="hidden" id="hdn_with_funding" name="with_funding">
+                <input type="hidden" id="hdn_with_other_legal_information" name="with_other_legal_information">
+                <input type="hidden" id="hdn_with_additional_info" name="with_additional_info">
+                <input type="hidden" id="hdn_with_keywords" name="with_keywords">
             </form>
 
             {% include 'learning_unit/blocks/form/search_form_reset.html' %}
